+++ src/util.hpp
@@ -19,7 +19,7 @@ inline const map_t& getMap(const gamestate_t& game) {
 	return game.level.map.find(game.world.level)->second;
 }
 
-inline void setView(glm::mat4& view, const glm::vec3& playerTranslation) {
+inline void setView(glm::imat4& view, const glm::vec3& playerTranslation) {
 	view = glm::translate(glm::ivec3(-playerTranslation.x + 16 * 4,
 									 -playerTranslation.y + 16 * 9,
 									 0));
